requirements.txt
flask==2.3.2
flask-login==0.6.2
werkzeug==2.3.7  
mysql-connector-python==8.1.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5